Clean up polymorphsystem (#14297)
