+++ Content.Server/Polymorph/Components/PolymorphedEntityComponent.cs
@@ -17,13 +17,13 @@ namespace Content.Server.Polymorph.Components
         /// The original entity that the player will revert back into
         /// </summary>
         [DataField("parent", required: true)]
-        public EntityUid Parent = new();
+        public EntityUid Parent;
 
         /// <summary>
         /// The amount of time that has passed since the entity was created
         /// used for tracking the duration
         /// </summary>
         [DataField("time")]
-        public float Time = 0;
+        public float Time;
     }
 }
